Rework config, make it easier to add new types and validate them accordanly, split up some models from files
