+++ src/models/NotifierService.ts
@@ -1,0 +1,3 @@
+export default interface NotifierService {
+  sendNotification(url: string, account: string): Promise<void>;
+}
