+++ src/notifications/index.ts
@@ -1,3 +1,16 @@
-export default interface NotificationService {
-  sendNotification(url: string, account: string): Promise<void>;
+import { NotificationType } from '../models/NotificationsType';
+import emailNotifier from './email';
+import telegramNotifier from './telegram';
+import localNotifier from './local';
+import config from '../config';
+import NotifierService from '../models/NotifierService';
+
+const notifiers: Record<NotificationType, NotifierService> = {
+  [NotificationType.EMAIL]: emailNotifier,
+  [NotificationType.TELEGRAM]: telegramNotifier,
+  [NotificationType.LOCAL]: localNotifier,
+};
+
+export default function getNotifier(): NotifierService {
+  return notifiers[config.notificationType];
 }
